01-getting-started/10-agent-observability-and-evaluation/datadog/adot/adot-stuff/update-requirements.txt
boto3
botocore
awscli
opensearch-py
requests-aws4auth
pyyaml
retrying
strands-agents
strands-agents-tools
langfuse
langchain-aws
pandas
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
requests
